chore: Modified  several something for 'token' package of infrastructure layer

+++ internal/infrastructure/token/claims.go
@@ -6,7 +6,7 @@ import (
 
 type Claims struct {
 	jwt.RegisteredClaims
-	UserID uint `json:"user_id"`
+	UserID uint64 `json:"user_id"`
 }
 
 // func (c Claims) Valid() error {
